requirements.txt
streamlit>=1.25.0
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
plotly>=5.8.0
opencv-python>=4.5.5
Pillow>=9.0.0
onnxruntime>=1.10.0